# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Contract Lifecycle Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contracts:
    get:
      operationId: listContracts
      tags: [contracts]
      summary: List contracts visible to the current user
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: direction
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of contracts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractListResponse"

    post:
      operationId: createContract
      tags: [contracts]
      summary: Create a new contract
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContractBody"
      responses:
        "201":
          description: Contract created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /contracts/{id}:
    get:
      operationId: getContract
      tags: [contracts]
      summary: Get a contract by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contract detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    patch:
      operationId: updateContract
      tags: [contracts]
      summary: Update a contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContractBody"
      responses:
        "200":
          description: Updated contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /contracts/{id}/submit:
    post:
      operationId: submitContract
      tags: [contracts]
      summary: Submit a contract for review (triggers AI screening)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contract submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /contracts/{id}/resubmit:
    post:
      operationId: resubmitContract
      tags: [contracts]
      summary: Resubmit a contract after edits
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contract resubmitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /contracts/{id}/rescreen:
    post:
      operationId: rescreenContract
      tags: [contracts]
      summary: Manually trigger AI re-screening
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Screening triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
  /contracts/{id}/approve:
    post:
      operationId: approveContract
      tags: [contracts]
      summary: Approve the current workflow stage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentBody"
      responses:
        "200":
          description: Stage approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /contracts/{id}/send-back:
    post:
      operationId: sendBackContract
      tags: [contracts]
      summary: Send a contract back to submitter for edits
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentBody"
      responses:
        "200":
          description: Sent back
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /contracts/{id}/comment:
    post:
      operationId: addComment
      tags: [contracts]
      summary: Add a comment to a contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentBody"
      responses:
        "201":
          description: Comment added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditEntry"
  /contracts/{id}/upload-document:
    post:
      operationId: uploadDocument
      tags: [contracts]
      summary: Upload a document for a contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadDocumentBody"
      responses:
        "200":
          description: Document uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /contracts/{id}/upload-executed:
    post:
      operationId: uploadExecutedDocument
      tags: [contracts]
      summary: Upload the fully executed contract document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadDocumentBody"
      responses:
        "200":
          description: Executed document uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /contracts/{id}/audit:
    get:
      operationId: getContractAudit
      tags: [contracts]
      summary: Get audit trail for a contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Audit trail
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditEntry"
  /contracts/{id}/screening:
    get:
      operationId: getContractScreening
      tags: [contracts]
      summary: Get AI screening results for a contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Screening results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScreeningResult"

  # Contract Types
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessMessage:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Contract:
      type: object
      properties:
        id:
          type: integer
        contractName:
          type: string
        contractTypeId:
          type: integer
        contractTypeName:
          type: ["string", "null"]
        direction:
          type: string
          enum: [buy, sell]
        status:
          type: string
        counterpartyName:
          type: string
        counterpartyAddress:
          type: ["string", "null"]
        contractValue:
          type: ["number", "null"]
        effectiveDate:
          type: ["string", "null"]
        expirationDate:
          type: ["string", "null"]
        autoRenewal:
          type: boolean
        noticePeriodDays:
          type: ["integer", "null"]
        description:
          type: ["string", "null"]
        department:
          type: ["string", "null"]
        driveFileId:
          type: ["string", "null"]
        executedDriveFileId:
          type: ["string", "null"]
        submittedById:
          type: integer
        submittedByName:
          type: ["string", "null"]
        currentStageId:
          type: ["integer", "null"]
        formData:
          type: ["object", "null"]
        aiRiskScore:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - contractName
        - contractTypeId
        - direction
        - status
        - counterpartyName
        - autoRenewal
        - submittedById
        - createdAt
        - updatedAt
    ContractDetail:
      allOf:
        - $ref: "#/components/schemas/Contract"
        - type: object
          properties:
            auditTrail:
              type: array
              items:
                $ref: "#/components/schemas/AuditEntry"
            screeningResult:
              $ref: "#/components/schemas/ScreeningResult"
            obligations:
              type: array
              items:
                $ref: "#/components/schemas/Obligation"
            currentStage:
              $ref: "#/components/schemas/WorkflowStage"
            workflowStages:
              type: array
              items:
                $ref: "#/components/schemas/WorkflowStage"
    ContractListResponse:
      type: object
      properties:
        contracts:
          type: array
          items:
            $ref: "#/components/schemas/Contract"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - contracts
        - total
        - page
        - limit
    CreateContractBody:
      type: object
      properties:
        contractName:
          type: string
        contractTypeId:
          type: integer
        direction:
          type: string
          enum: [buy, sell]
        counterpartyName:
          type: string
        counterpartyAddress:
          type: ["string", "null"]
        contractValue:
          type: ["number", "null"]
        effectiveDate:
          type: ["string", "null"]
        expirationDate:
          type: ["string", "null"]
        autoRenewal:
          type: boolean
        noticePeriodDays:
          type: ["integer", "null"]
        description:
          type: ["string", "null"]
        department:
          type: ["string", "null"]
        driveFileId:
          type: ["string", "null"]
        formData:
          type: ["object", "null"]
      required:
        - contractName
        - contractTypeId
        - direction
        - counterpartyName
        - autoRenewal
...
    UploadDocumentBody:
      type: object
      properties:
        driveFileId:
          type: string
        fileName:
          type: string
      required:
        - driveFileId
        - fileName
    CommentBody:
      type: object
      properties:
        comment:
          type: string
      required:
        - comment
...